requirements.txt
numpy
scipy
boolean.py
jupyter
notebook
networkx
maboss
